score/requirements.txt
click==8.1.7
grpcio==1.68.1
grpcio-reflection==1.68.1
grpcio-tools==1.68.1
h11==0.14.0
protobuf==5.29.2
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
python-dotenv==1.0.1
setuptools==75.5.0
typing_extensions==4.12.2
uvicorn==0.33.0